Moved plotting functions to seperate file
